HUBComponent: Make `componentIdentifier` nullable
We don’t want to require API users of the upcoming `HUBViewModelBuilder`
to have to up-front specify which component that a certain model should
be rendered using. Therefor, make `componentIdentifier` nullable and
update the documentation.

The upcoming `HUBViewController` will use its `HUBComponentFallbackHandler`
to deal with nil or unknown component identifiers.