Add JSON paths; the core of the Hub Framework JSON parsing mechanism
The Hub Framework uses a path-based approach to parsing JSON, where each
path is defined as a sequence of atomic operations. The reasoning behind
this approach, rather than using a dictionary of static strings (for example)
is that the framework needs to be flexible enough to support a vast range
of JSON formats and backend services.

This commit contains the core of the parsing mechanism, the immutable and
mutable version of `HUBJSONPath` that enables the API user to define paths.

Most of the code has been lifted over from the HubFeature in client-ios, but
the API has been heavily simplified and the strong type safety previously
employed for enum values has been scaled back in favor of flexibility.