Add `HUBComponentModel.targetInitialViewModel`
This property will enable API users to construct a tree of views up-front
(either through JSON or the local content provider API). This is a more simple,
and way more powerful, implementation of what is the HubFeature is called
`targetViewTitle`, `targetViewHeaderStyle` etc.

The purpose of this API is to enable the creation of “skeleton” views, and
not have all views be completely black until the initial data set has been
loaded, providing the user with a better experience.