HUBComponent: Use UIViewController-style loadView/view pattern
This commit changes the `-loadView` method of a `HUBComponent` to return
`void` and instead require the component author to assign a new `view`
property. This is done to make implementing components easier (no custom
view properties required) and also makes the pattern of implementing a
`HUBComponent` similar to how `UIViewController` works.