Add HUBViewModelLoaderFactory
Used to create view model loaders both inside & outside of the Hub Framework
(used externally for the external data API)

An instance of `HUBViewModelLoaderFactory` is created in `HUBManager`.

`HUBContentProviderFactoryMock` has also been updated to not always return
nil, and instead enable a test author to test the remote/local content providers
that it creates.