Change component creation to use a factory.
Use the factory pattern instead of using component instances as
prototypes for creating new instances. This allows for a cleaner
separation of creating new components from the logic of the components
while still supporting the dependency injection pattern for components.

This change also move the fallback logic from a
HUBComponentFallbackHandler into the component factories, allowing for
less centralized fallback handling.