Fix build errors when compiling outside of client-ios
Since the config that the Hub Framework uses when built outside of
client-ios is a lot stricter than the one used in client-ios, some
build errors need to be fixed.

Fixes include adding temporary variables for `nullable` parameters,
as well as adding a macro for checking what version of the iOS SDK
we’re building against.