Refactor `HUBComponentIdentifier` for easier component creation code
- Make `HUBComponentIdentifier` private to the framework and change its
`componentNamespace` property to be `nonnull`.
- API users now deal with `componentNamespace` and `componentName`
separately.
- If a content provider fails to set a component name for a component,
that component will not be rendered by the Hub Framework.
- When setting up `HUBManager`, a default namespace and a fallback name
are now passed, instead of an identifier.

This commit will make the logic in `HUBComponentRegistry` and
`HUBViewController` trivial, by reducing the number of combinations that
a component identifier can have. It also reduces the likelihood of
discrepancies between the UI and the component factories.