Implement image loading
This commit implements an image loading mechanism for components displayed
in a HUBViewController. It does so using a contextual object that tracks
image downloads, to make sure that the right image gets mapped to the
right component.

One tweak to HUBViewController´s API that needed to be made in order to
test the image loading mechanism is exposing its underlying collection view
as readonly.