Add dynamic, trait-based layout system & custom collection view layout
This commit implements a new, dynamic layout system for the Hub Framework.
The system is based on traits - `HUBComponentLayoutTrait` - that each define
a characteristic of a component. Using these traits, an implementation of
`HUBComponentLayoutManager` can make deterministic decisions on how to lay
components out on the screen, in relation to each other and to the content
edges of a view.

Using these traits, and the decisions made by `HUBComponentLayoutManager`,
a custom `UICollectionViewLayout` is computed (up-front for maximal scroll
performance), that translates the layout traits given by components into
absolute frames that then can be rendered by `UICollectionView`.

The reasoning behind this approach is that different components will need
different margins, depending on the context they are used in. So a standard
`UICollectionViewFlowLayout` won’t do the trick, since it’s based on sections.
We also don’t want components themselves to have to handle their layout, as
it extends the area of responsibility of a component too much.

Using layout traits, a component doesn’t need to be aware of its surroundings,
or of other component types, but can instead declare its desired layout in
a very flexible way.

This change also moves all layout computations out of the view controller.