Refactor `HUBContentProviderFactory` into local & remote + add URL resolver
This commit refactors the `HUBContentProviderFactory` API into 2 separate
protocols - `HUBRemoteContentProviderFactory` and `HUBLocalContentProviderFactory`.

This is the way things were working in the Hub Framework 1.0, and initially
the idea was to keep the API more simple by using a unified factory, rather than
two separate.

However, when starting to port the `HUBRemoteContentURLResolver` API from
1.0 to 2.0, it became clear that the factories need to be separate in order for
a feature to combine using the URL resolver API (which most do today), with using
local content.

In order to not put networking code in the Hub Framework (to be able to use
external libraries, such as SPTDataLoader), a `HUBDefaultRemoteContentProviderFactory`
is introduced. This factory will create remote content providers based on URL resolvers,
and will be implemented in client-ios.