Enable components to create child components through the framework
This change renames `HUBComponentChildEventHandler` to `HUBComponentChildDelegate`,
and adds an API to it that enables components to easily create child component
instances through the framework, instead of having to do it manually.

Also extract component view loading code to `HUBComponentLoadViewIfNeeded`, and place
it in `HUBUtilities`, since a component’s view can now be loaded from several places.