HUBComponent: Add API to get notified of view resizes
Since the size returned by a HUBComponent’s -preferredSizeForDisplayingModel…
is not guaranteed to be its final size (since it may be resized due to its
layout traits, screen rotations, etc), we need an override point where the
component can react to its size being changed.

This is especially important for components that manage children or other
types of subviews, where frames may need to be recalculated.

This new API is anticipated to be used by a broad range of components, so
that’s why it’s introduced to the main HUBComponent protocol, rather than as
an extension protocol.