HUBViewModelBuilder: Don’t override existing data when adding JSON data
Before this change, when adding data from JSON to a view model builder,
all the previous data would be overridden (even if the JSON didn’t contain
a value for that specific piece of data).

While this was fine when local content was always loaded after remote,
with the new content provider chaining API, JSON & local content need to be
treated with equal weight.

After this change, only non-`nil` JSON values are added to the builder.
Custom data is merged between data added in code & through JSON.

The same change will also be performed for other builders in subsequent commits.