Replace root View URIs + HUBViewURIQualifier with HUBViewURIPredicate
This change removes the use of “root view URIs” or a “view URI qualifier”
when setting up a feature with the Hub Framework, and instead merges the
two concepts into `HUBViewURIPredicate`, a more flexible and powerful way
to determine what view URIs are part of a given feature.

HUBViewURIPredicates work just like NSPredicate (in fact, you can initialize
the former with the latter), but with an API dedicated to evaluating view
URIs. It has convenience constructors for all common use cases, such as
prefixing and using constant view URIs.