HUBViewController: Replace header visibility API with view model based one
This change removes the current view controller delegate API that notifies
the delegate when a Hub Framework view controller changed the visibility of
a header view component, and replaces it with one that notifies the delegate
when a new view model started being used.

This enables `client-ios` to perform operations on the container view controller
based on the current view model (such as setting nav bar titles & preparing for
state restore).