Add serialization support to all all core models
This change adds APIs to serialize `HUBViewModel`, `HUBComponentModel`
and `HUBComponentImageData`. This is done by using the default Hub Framework
JSON schema.

Serialization has many applications, such as caching a view model, implementing
state restore, or in automated UI tests.

To not duplicate identifiers between the JSON parsing code and serialization
code, all keys used for JSON are moved to `HUBJSONKeys.h`. The creation of
a default `NSDateFormatter` is also moved to `HUBUtilities.h`.