Add HUBContentReloadPolicy
This API is a port of `SPTHubReloadPolicy`, which is used by features that
wish to create their own logic around when a view’s content should be reloaded.

A new parameter to `HUBManager` is also introduced, where a default content
reload policy should be passed. `client-ios` will setup a default reload policy
using this, so that every feature doesn’t have to declare a reload policy if
they don’t want to.