Replace all uses of NS_DESIGNATED_INITIALIZER with HUB_DESIGNATED_INITIALIZER
This new macro adds the `NS_DESIGNATED_INITIALIZER` annotation to a designated
initializer, as well as making the default Foundation initializers (init + new)
unavailable, so that unavailable annotations don’t have to be added for every
single header file.

The result is a little bit of extra compiler magic, but a lot more consistency
and cleaner interfaces.