Add HUBComponentDefaults that encapsulates default values for component models
This change refactors the way that default values for component namespaces &
names are passed around in the Hub Framework. Before, string values were passed
around, but with this change they will be encapsulated in a `HUBComponentDefaults`
object.

The reason for this is that new default values can easily be added, instead of
having to go through the framework and adding parameters for new values. It also
makes it easier to keep consistency within the framework.

This change also introduces a slight API change. Before, not setting a component
name on a `HUBComponentModelBuilder` caused the builder not to generate a model.
This behavior has proven (in workshops) to be quite confusing, as there is no
real indication (except for documentation) that the field was required. Now, the
default value defined by `HUBComponentDefaults` will be used in case no explicit
component name was defined, allowing us to remove `nullable` from the component
name property.