Add component category & new fallback handler API
This change introduces a new `componentCategory` property on `HUBComponentModel`
(& its builder), that together with the new `HUBComponentFallbackHandler` API can
be used to enable more intelligent fallback for components.

Previously, in case a component couldn’t be resolved for a given identifier, the
Hub Framework would use a single identifier as a fallback. While this ensured that
_something_ will always be rendered, it didn’t enable feature developers to specify
what category of component to fall back to.

Now, with this change, each component model can declare a category, and the fallback
handler will be asked to return a fallback component for that category. In client-ios,
we can add categories for our commonly used component paradigms, such as “row”, “card”,
“carousel”, etc.