Forward view controller background color to collection view
This change introduces a `HUBContainerView`, that enables the Hub Framework
to guarantee that the same background color as being assigned to a view
controller’s view also gets assigned to the collection view that contains all
component views.

This enables client-ios to assign the background color of a Hub view controller
without being aware of its internal view hierarchy.