Add HUBIconImageResolver for easy icon image handling
This change introduces the `HUBIconImageResolver` API, which is a new
dependency for `HUBManager`. Using the icon image resolver, components
can now easily retrieve a renderable image for either component icons
or placeholder icons.

Before, each component was responsible for resolving icons manually
from identifiers, which renders a lot of code duplication and margin
for errors.

With this change, we can now have a centralized place where all icon
images are resolved, making it easier to create components that utilize
icons.