Add HUBComponentViewObserver sub-protocol with -viewWillAppear API
This change introduces a new `HUBComponentViewObserver` protocol that enables
a component to observe its view for changes. This protocol has a new
`-viewWillAppear` method that can be used to trigger animations and reset
selection state, as well as `-viewDidResize` (which has been moved from the
main `HUBComponent` protocol - reducing boilerplate).

The implementation uses different techniques on iOS8+ & on iOS7 due to that
`collectionView:willDisplayCell:forItemAtIndexPath` was introduced in iOS8.

To be able to unit test this fully, the current `UIDevice` is injected into
`HUBViewControllerImplementation`, so that it may be mocked in unit tests.

`HUBUtilities` is now also `nullability` audited, which required some misc
changes wherever its utility functions were being used.