Refactor content operation chain to be queue-based
This change refactors the way content operations are performed when building a view model.
Previously, `HUBViewModelLoader` kept track of the current index of a content operation, and
based a lot of decisions on that.

With this new approach, content operations are wrapped (`HUBContentOperationWrapper`), and
are put in a queue whenever triggered for rescheduling. That way the content loading chain
becomes a lot easier to reason about, easier to test, and more predictable.

This change also removes the need for `HUBContentOperationMode`, as the queue-based approach
doesn’t require us to keep track of if a content operation is sync or async. It also requires
content operations from now on to explicitly reschedule themselves instead of calling their
delegate method multiple times.