Add support for prepended & appended content operation factories
This change adds two new arguments to `HUBManager`’s initializer, that
enables an application using the Hub Framework to supply prepended and
appended content operation factories.

The use case for this is to be able to define global pre- and post
processing content operations that can be used to create loading spinners,
show error- or info views, annotate playback/highlighting, etc.