Add support for overlay components
This change introduces a 3rd class of components (in addition to “header”
and “body”) - “overlay”. Just like Header components, Overlay components
are rendered above the content that is being displayed in the body, with
the key difference that they are always centered in the middle of the view.

In client-ios, we will use overlay components to implement loading spinners,
info views, etc.