Add support for multiple overlay components in a view
This change enables a view to have multiple, stacked, overlay components.
Previously, only a single overlay component could exist per view, and if
that component’s model was overwritten, the overlay would be replaced.

This change ensures a higher level or predictability for overlays, since
an overlay may now be tracked using its model identifier, rather than
assuming that the overlay that is currently displayed is the one that was
created before.

This change also includes some slight header cleanup, with grouping of
similar methods.