HUBManager: Make defaultContentReloadPolicy an optional parameter
This is part of an effort to reduce the boilerplate required to setup
an instance of the Hub Framework in a new application. Even though it’s
recommended to use a default content reload policy, it’s not strictly
required - so the parameter is now optional - making a reload policy
optional overall.

A feature not using a reload policy will always be reloaded.