JSON: Allow invalid elements in an array (just don’t include them)
This change makes the JSON parsing of the Hub Framework always return
an array when asked to do so, instead of returning `nil` when an invalid
element was encountered.

This is done to reduce the threshold of failure, but also to be able to
implement fallback logic using combined paths, as one path might fail -
which shouldn’t cause the operation itself to fail.