HUBViewModel: Replace entity identifier with view URI
Entity identifier is no longer needed for instrumentation of client-ios,
and doesn’t make much sense outside of the Spotify app either. View URI
on the other hand, is a more generic, core concept of the framework and
can have similar purposes.

Exposing the view URI as part of the view model also enables us to easier
pass that information to various consumers of the framework.