Support saving & restoring UI state for components
This change introduces a new sub-protocol for `HUBComponent`, which extends
it with functionality for managing UI state (such as scrolling positions,
selection state, etc - that is not expressed in the model layer).

The new protocol; `HUBComponentWithRestorableUIState` is then used by the
Hub Framework and a new `HUBComponentUIStateManager` class to save, restore
and manage the UI state of components.

Since this is a new protocol, the functionality is opt in.