Support deep hierarchies of components, rather than just 2 levels
This change addresses issues with using component hierarchies with more
than 2 levels, that is children of children of a component. Before this
change, any children who also had children of their own wouldn’t get assigned
delegates etc, because internally a `HUBComponentWrapper` wouldn’t be created
for them.

This fix makes sure that for each child component that gets created, a component
wrapper is also created.

This is not the final solution to this problem, which involves a more thorough
rethinking of the component management system, where each level in the hierarchy
is independently managed. It should, however, solve our most immediate needs, and
unblock teams who have started to work on these type of hierarchies.