Vertical paging scroll behavior: Remove trailing “f”s
Adding trailing “fs” to a numeric value makes it a float, however, these
values are being used for `CGFloat` values which are typedef’d as doubles
on 64 bit systems, making the “f” incorrect. Removing it is the “safe”
option since it covers both the 32 bit and 64 bit use case.

See CoreGraphics/CGBase.h for more information.