Add basic implementation of “Smart reload”
With this change, a Hub view controller’s UICollectionView is only
reloaded if the view model was modified since the last rendering pass.

This is done by overriding the setters of all builder properties (using
a macro, to avoid code duplication) and keeping track of when they change.

As mentioned in the title, this is a very basic implementation of what
was outlined in https://ghe.spotify.net/iOS/HubFramework/issues/234, one
could imagine a more thorough implementation that checks the modification
status of each individual component.

However, for the time being, this should fix the glitches and image
problems we’ve been seeing for views that are rapidly reloaded without
mutating their view model.