Ensure that collection view is always first subview of view controller
This fixes another regression caused by moving the collection view creation
code to -viewWillAppear. Before, there was a guarantee that the collection
view would be the first view created, and thus always be the first subview
(and have the lowest z-index) of all subviews.

This was causing a problem with overlay components being added before the
view will appear (like, in the Spotify app, when we’re adding a progress
indicator as part of the initial view model). Those overlay components
would be added below the collection view, and thus not be visible.