View model loader: Reset state when connectivity state changes
This change makes HUBViewModelLoader completely reset its state whenever
it is reloading due to a change in connectivity state. The reason we want
to do this is to be able to use the initial view model (which in the case
of the Spotify app contains the loading indicator).

This change also changes the nullability of `initialViewModel` to `nonnull`,
since these days its always computed.