View model builder: Add API to enumerate all component model builders
This new API can be used to sequentially inspect the component content of
a view model builder, which becomes very useful in content operations that
behave differently depending on the state of a view.

A concrete example; in the Spotify app, we’re add a loading indicator in
form of an overlay component. Whenever a content loading chain comes to an
end (at an appended content operation), we want to check whether a view model
contains anything _else_ than just the loading indicator.

This API makes that possible in a very efficient and clean manner, rather than
having to check multiple states and compare them.

This new API is also used internally in the view model builder to gather all
body component model builders, instead of using a classic for loop.