ViewController: Create collection view if needed in loadView & viewWillAppear
We want to be able to make the guarantee that the collection view is available
in -viewDidLoad, since parts of the Spotify infrastructure relies on content
scroll views being available at that point. It also makes for a nicer and
more predictable API.

A while back, the creation of collection views was moved to -viewWillAppear,
to make the view controller recover from a low memory situation in a correct
way. That however, has caused several regressions - so this patch gives us
the best of both worlds.