Move component selection handling code into selection handler wrapper
This change is targeted at isolating the code for performing component
selection into a new class `HUBComponentSelectionHandlerWrapper`. As the
name implies, this class also wraps any custom selection handler that a
feature defined - making it a single entry point for the view controller
to perform selection handling.

The motivation for performing this change is firstly to create a better
separation of concerns when it comes to the view controller, but also to
prepare for the upcoming `HUBManager` (https://ghe.spotify.net/iOS/HubFramework/issues/250),
which will use this new selection handler instead of the view controller.