Fix view controller crash on significant content update
This patch fixes a crash that would occur if a Hub view model was significantly
updated (that is, all content removed, or several new component models added)
since the last rendering pass.

The problem was that the collection view’s underlying data hadn’t been updated
at the point in time where the layout was changed, which would cause an internal
UIKit error (which would manifest itself as a `EXEC_BAD_ACCESS` in the app).

The fix is to call `reloadData` before updating the layout, which then in turn
surfaced another problem, that `collectionView:didEndDisplayingCell:…` had a
data inconsistency problem, which has now also been fixed.