View controller: Remove unreused overlay components
This patch fixes a bug where overlay components that weren’t reused would
still be rendered in the view. This was caused by separate code paths being
executed when an overlay component model was completely removed versus when
the component id was simply changed.

This bug would manifest itself in the Spotify app when slow connections would
cause an offline info view and progress indicator to be displayed at the same
time. When the content would finally load, the progress indicator would be
removed, but the info view would remain in the view - blocking the additional
content.