Make content loading chain truly declarative with view model snapshotting
This is is a pretty small but very significant change in how the content loading
chain works, and how a `HUBViewModelLoader` executes it, especially in terms of
how it handles rescheduled content operations.

Currently, a content operation can receive 2 kinds of view model builders as input:

1. When it’s being executed as part of the initial content loading chain, it will
receive a view model builder that is the product of all *previous* content operation
mutations. That is, the input for `C` is a view model containing mutations from `A`
and `B`.

2. When it’s being executed because it was asked to be rescheduled, it will receive 
a view model builder that contains all the *current* content of the view. That is, the
input for `C` is a view model contain mutations from the whole last content loading
chain, including mutations from `A`, `B` *and* `C`.

This basically creates a need for content operations to keep state, as they cannot be
guaranteed in what context they will be executed. Should a content operation add all of
its content, remove any previous content or just add new content? It becomes really hard
to tell.

With this change, 1 & 2 above get unified. Now, a content operation will always receive
a view model builder that is the product of the preceding content operations, nothing else.
That means that `C` will never receive a view model builder that already contains `C`
content (only `A` and `B` content, always).

This enables content operations to be truly declarative, and always setting up their content
from a “blank slate” instead of having to keep state. So the following code:

```
if (isOffline) {
    [viewModelBuilder builderForBodyComponentModelWithIdentifier:@“id”].title = @“You’re offline!”;
} else {
    [viewModelBuilder removeBuilderForBodyComponentModelWithIdentifier:@“id”];
}
```

Can now be reduced to the following:

```
if (isOffline) {
    [viewModelBuilder builderForBodyComponentModelWithIdentifier:@“id”].title = @“You’re offline!”;
}
```

This change has also been documented as part of the “Content programming guide”.