Components: Make sure -loadView is always called before -configureViewWithModel
This patch ensures (through an assertion) that the framework will always call
-loadView before it calls -configureViewWithModel on a component. Components
should be able to make the assumption that their view will be fully loaded
before they are asked to perform data binding (since that actually requires
a view to be done) - and this new mechanism ensures that correctness.

This patch also fixes 2 cases of the inconsistency mentioned above, which were
identified through the new assertion.