Refactor `HUBScrollBehavior` into `HUBViewControllerScrollHandler`
This change removes `HUBScrollBehavior` and its implementations and replaces
it with a new public protocol - `HUBViewControllerScrollHandler`.

The reason for this refactor is to move all things related to custom scrolling
behaviors and event handling out of the Hub Framework itself. The new API
follows the design of similar APIs (such as `HUBComponentSelectionHandler`) and
is registered in a similar way (through `HUBFeatureRegistry`).

This change also makes the separation of concerns a lot clearer, as the layout
code doesn’t have to be aware of scrolling behaviors, and content insets is now
done through actual content insets instead of manipulating component margins.