Move component target information into its own object
This change moves all component target information (targetURL + targetInitialViewModel)
into a new `HUBComponentTarget` object, instead of being attached directly onto
`HUBComponentModel` instances. This is done both to make the model structure match the
default JSON schema, but also to prepare for the upcoming Action API, which will add
action identifiers to targets, further warranting its own object.

In this patch `HUBViewModelImplementation` is also made a subclass of `HUBAutoEquatable`,
to make it able to get automatic equality checks, since we want to be able to auto-equate
a `HUBComponentTarget` instance.