Move “featureIdentifier” from View Model to View Controller & make immutable
This change resolves https://ghe.spotify.net/iOS/HubFramework/issues/271, and
makes the iOS implementation of the Hub Framework consistent with Android -
in that once a feature identifier has been specified when setting up a feature
with the framework, it cannot be changed.

This removes the need of the View Model & View Model Builder (and by extension
Component Model Builder - since they can have Initial View Models) to know
about feature identifiers - which leads to a lot of code being removed (yay!).

The feature identifier of a view controller is now exposed as part of the
`HUBViewController` API instead, where it can be consumed by, for example,
the Spotify logging system.