Add initial infrastructure for registering actions
This is the beginning of the Hub Framework Action API, which will provide an easy
way to extend the framework with functionality that should be performed in response
to an event.

Actions will be performable both from a selection event, but also manually by component 
implementations.

Actions follow the same architecture as components, in that they are registered with
the framework using a factory, for a given namespace. This way we can protect ourselves
against names/identifiers clashing between features.

This change includes the following:
- Add `HUBAction`, the protocol used to define actions
- Add `HUBActionFactory`, the protocol used to define action factories
- Add `HUBActionContext`, a contextual object used for actions. Will replace
  `HUBSelectionContext` in an upcoming commit, as selections will also be actions.
- Add `HUBActionRegistry`, similar to the component registry, this enables factories
  to be integrated with the framework.