Replace selection handling with actions
This change replaces all selection handling code with actions. This means that both
internally within the framework, and as part of the selection handling API - actions
are used both for the default selection events, and for custom events as well.

The following API changes are performed:
- `HUBComponentSelectionHandler` is replaced by `HUBActionHandler`.
- `HUBComponentSelectionContext` is removed (`HUBActionContext` is now used instead).
- A feature can now define an action handler instead of a selection handler.
- A default action handler can be supplied when setting up `HUBManager`. This follows
  same architecture as for content reload policies.

In a subsequent commit, the ability for components to trigger actions themselves,
as well for content operations to be able to respond to actions will be added.