HUBJSONPath: Return NSArray<id> from -valuesFromJSONDictionary:
Previously NSArray<NSObject *> was returned, meaning that casting had to
be used to get the correct value. This isn’t actually necessary in this
case, since HUBJSONPath actually includes a type-checking mechanism, meaning
that the casting just adds additional clutter that doesn’t improve the
type safety - so let’s remove it.