Add HUBComponentModel API for retrieving a child for a given identifier
This new API provides a O(1) complex operation to search for a child with
a given identifier, instead of having to iterate through all children and
comparing their identifiers.

This also opens up for components to be able to handle children in a new,
more dynamic way - since they are no longer constrained to creating children
based on index - and can instead opt to pick child models based on their
index.

For example; you could create a component with a title & button as child
components, in this way now:

```
Component {
   “title”: TitleComponent
   “button”: ButtonComponent
}
```

Where before, you’d have to either force the child components to be in a
certain sequence, or have to code around it:

```
Component [
   TitleComponent,
   ButtonComponent
]
```