Modernize child component model builder APIs
This change migrates the APIs that are used to create and manage child
component model builders to the same naming convention as was recently
introduced for child component models on `HUBComponentModel`.

The changes are made for API clarity and simplification, to remove words
that are unnecessary for convening the intent of the APIs.

Changes:

- `builderForChildComponentModelWithIdentifier` -> `builderForChildWithIdentifier`
- `builderExistsForChildComponentModelWithIdentifier` -> `builderExistsForChildWithIdentifier`
- `removeBuilderForChildComponentModelWithIdentifier` -> `removeBuilderForChildWithIdentifier`
- `allChildComponentModelBuilders` -> `allChildBuilders`
- `removeAllChildComponentModelBuilders` -> `removeAllChildBuilders`
- `childComponentModelDictionariesPath` -> `childDictionariesPath`