Add error snapshotting for content operations
This change adds snapshotting functionality for errors, just like was introduced
for view model builders a while back. Snapshotting enables content operations to
assume that they always will be executed from the same previous state when
rescheduled, so that they don’t have to keep their own states & caches as much.

With this change, any content operation that gets rescheduled will be passed the
same error as it was passed the last time it was run. This prevents subtle bugs
that would occur because a content operation would reschedule itself - but now
without an error it would yield a completely different (unwanted) result.