Support selection for both collection- & table view cell components
This patch fixes a problem where components that which view was implemented
using `UITableViewCell` wouldn’t be selectable, because of the way UIKit
internally handles touch events for cells.

It looks like UIKit performs an `[isKindOfClass:[UITableViewCell class]]` check
(have “confirmed” this using some thorough breakpointing), and then looks for a
parent `UITableView` in order to handle the touch (which it won’t find, resulting
in a prevented touch event).

We previously “solved” this for `UICollectionViewCells`, by setting
`userInteractionEnabled = NO` on them. That always felt hacky, but also disabled
components implemented using `UICollectionViewCell` to do any custom touch handling
inside of them (which became even more obvious with the new Actions API).

With this solution, we override `pointInside:withEvent:` to “trick” UIKit that
no cell has been selected, while manually forwarding all touch events to the
component’s view. The result is that the cells can remain user intractable, without
disrupting the selection handling of the root `UICollectionView`. It works for both `UITableViewCell`- and `UICollectionViewCell`-based components.