Add default image loader implementation
This image loader is very simple, but provides a way for applications that
do not define their own image loader to still get image loading support from
the Hub Framework.

Previously, supplying `nil` as `imageLoaderFactory` when setting up `HUBManager`
caused the framework to be incapable of loading images. With this change, the
barrier for entry becomes lower to start using the framework, while still enabling
images to be downloaded.