Introduce component model groups
Grouping enables components to support more sophisticated hierarchies of
child components. For example, in a header component, grouping can be used
to differentiate an action button child component from contents of a page
component.

In the Spotify app, we can for example model the entity header view like this:

```
EntityHeaderComponent
	- buttonComponent (group = “button”)
	- pageComponentA (group = “pages”)
	- pageComponentB (group = “pages”)
```