Adapt overlay component center point to content inset + keyboard
With this change, a `HUBViewControllerImplementation` instance starts observing
when the keyboard is shown & hidden, for as long as its visible on the screen.
Whenever the keyboard is shown, any displayed overlay components are adjusted
to be centered in the area between the navbar & the keyboard.

This solves a problem that became apparent in the demo app, where overlay
components would appear misaligned when a search bar became focused and the
keyboard appeared.

(Also make search bar loose focus when search button is clicked, made it a
lot simpler to test this change).