Enable content operations to call their delegate on a background queue
With this change, content operations are now free to call their delegate’s
methods on any queue, and the Hub Framework will automatically dispatch to
the main queue (if needed). To do this, a new `HUBPerformOnMainQueue()`
function is introduced in HUBUtilities, which is now also used for image
loader callbacks in the view controller.