[Demo app] Add “Really long list” feature
This new feature adds 10,000 row components to a view controller, and will
act as a demonstration of how well the Hub Framework handles a large number
of components, thanks to us using Collection View and view reuse.

It also provides a nice playground for testing out performance improvements
such as improving the speed of our diffing algorithm. In the future, we could
even run UI performance tests against this view to make sure we’re always
rendering in a smooth 60 FPS irregardless of how may components are used.