[Demo app] Clean up GitHub Search code + add comments
The GitHub Search content operation code was a bit hackily implemented
(I blame it on lack of coffee!), and since it’s a bit complicated, it
also suffered from not being commented.

Both of these issues have now been fixed, it should now be a lot simpler
to follow as it has a more streamlined implementation and comments explaining
why we’re doing what we’re doing.