Add APIs for getting number of body/overlay component builders in a view model builder
Previously, you could get the number of body or overlay component model builders
contained within a view model builder by enumerating them. This was a bit wasteful
if all you’re looking for is the count, so these new APIs are more efficient for
that use case.

Also do some slight reordering of methods in the implementation to match the way
the header is declared.