HUBManager: Add convenience initializer
Since we want to support a lot of customization options for the Hub Framework
(we also need them for the Spotify app), the initializer for HUBManager is
quite large. However, most of its arguments are optional, so with this change
a convenience initializer is added that only takes the required arguments.

Also update the demo app to using this new initializer.